move type aggregation
