+++ vestapol/destinations/gcp_destination.py
@@ -8,7 +8,7 @@ from vestapol.destinations import base_destination
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from vestapol.typing.resources import ResourceTypes
+    from vestapol.web_resources import ResourceTypes
 
 logger = logging.getLogger(__name__)
